# specs/001-course-assistant-mvp/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DOU-Synapse API",
    "description": "Ders materyaline bağlı, kaynaklı cevap üreten ders asistanının HTTP yüzeyi. Her cevap dayandığı sayfayla döner; dayanak yoksa uç cevap değil gerekçeli ret döndürür. Ders izolasyonu iki katmanda zorlanır: sunucu tarafında üyelik doğrulaması ve aynı oturumda PostgreSQL satır düzeyi güvenlik.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "List Admin Users",
        "description": "Kullanıcı dizinini URL'ye ad/e-posta bırakmadan sorgula.",
        "operationId": "list_admin_users_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserQueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Admin Courses",
        "operationId": "list_admin_courses_admin_courses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCourseListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/requests": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Admin Requests",
        "operationId": "list_admin_requests_admin_requests_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "route",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Route"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnswerStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRequestListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/ingestion": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Admin Ingestion",
        "operationId": "list_admin_ingestion_admin_ingestion_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIngestionListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List My Courses",
        "description": "Kullanıcının aktif üyeliği olan dersler.",
        "operationId": "list_my_courses_courses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_CourseWithRole_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course",
        "description": "Ders oluşturur ve oluşturana aynı işlemde eğitmen üyeliği verir.\n\nİki kayıt tek işlemde yazılır; aksi halde sahipsiz (kimsenin göremediği) bir ders\nkalabilirdi.",
        "operationId": "create_course_courses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseWithRole"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminCourseItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "creator_name": {
            "type": "string",
            "title": "Creator Name"
          },
          "active_member_count": {
            "type": "integer",
            "title": "Active Member Count"
          },
          "documents_total": {
            "type": "integer",
            "title": "Documents Total"
          },
          "documents_failed": {
            "type": "integer",
            "title": "Documents Failed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "created_at",
          "creator_name",
          "active_member_count",
          "documents_total",
          "documents_failed"
        ],
        "title": "AdminCourseItem"
      },
      "AdminCourseListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminCourseItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminCourseListOut"
      },
      "AdminIngestionItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "course_id",
          "course_code",
          "status",
          "attempt_count",
          "started_at",
          "completed_at",
          "created_at"
        ],
        "title": "AdminIngestionItem"
      },
      "AdminIngestionListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminIngestionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminIngestionListOut"
      },
...
      "AdminRequestItem": {
        "properties": {
          "log_id": {
            "type": "string",
            "format": "uuid",
            "title": "Log Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "route": {
            "type": "string",
            "title": "Route"
          },
          "mode": {
            "$ref": "#/components/schemas/ChatMode"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "http_status": {
            "type": "integer",
            "title": "Http Status"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "course_id",
          "course_code",
          "route",
          "mode",
          "status",
          "http_status",
          "latency_ms",
          "token_count",
          "cache_hit",
          "created_at"
        ],
        "title": "AdminRequestItem"
      },
      "AdminRequestListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminRequestItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminRequestListOut"
      },
      "AdminUserItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "masked_email": {
            "type": "string",
            "title": "Masked Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_platform_admin": {
            "type": "boolean",
            "title": "Is Platform Admin"
          },
          "active_course_count": {
            "type": "integer",
            "title": "Active Course Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "masked_email",
          "full_name",
          "created_at",
          "is_platform_admin",
          "active_course_count"
        ],
        "title": "AdminUserItem"
      },
      "AdminUserListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminUserListOut"
      },
      "AdminUserQueryIn": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 25
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          },
          "search": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Search"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminUserQueryIn",
        "description": "PII içerebilecek aramayı URL yerine doğrulanan gövdede taşır."
      },
...
      "AnswerStatus": {
        "type": "string",
        "enum": [
          "answered",
          "insufficient_context",
          "out_of_scope",
          "budget_exhausted"
        ],
        "title": "AnswerStatus",
        "description": "Cevabın dört meşru sonucu. Üçü 'cevap yok' ama SEBEPLERİ farklı.\n\nKarıştırılmamalı: `insufficient_context` \"materyalde var olabilir ama kanıt\nzayıf\", `out_of_scope` \"bu ders bu konuyu hiç kapsamıyor\". Kullanıcıya\nverilen metin de ölçüm kategorisi de farklıdır (SC-005 yalnız ikincisini ölçer)."
      },
...
      "ChatMode": {
        "type": "string",
        "enum": [
          "qa",
          "socratic",
          "exam"
        ],
        "title": "ChatMode"
      },
...
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Code",
            "examples": [
              "COME301"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title",
            "examples": [
              "İşletim Sistemleri"
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "title"
        ],
        "title": "CourseCreate"
      },
      "CourseWithRole": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "created_at",
          "role"
        ],
        "title": "CourseWithRole",
        "description": "Listelemede kullanıcının o dersteki rolü de döner."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
...
      "MembershipRole": {
        "type": "string",
        "enum": [
          "instructor",
          "student"
        ],
        "title": "MembershipRole"
      },
...
      "PageOut_CourseWithRole_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseWithRole"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PageOut[CourseWithRole]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}